# docs/swagger.yaml
basePath: /api
definitions:
  models.AuctionPayload:
    properties:
      endTime:
        type: string
      nftAddress:
        type: string
      paymentToken:
        type: string
      startPrice:
        type: string
      startTime:
        type: string
      tokenId:
        type: integer
    required:
    - endTime
    - nftAddress
    - startPrice
    - startTime
    - tokenId
    type: object
...
  response.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A RESTful API for NFT auction marketplace using Gin and GORM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: My Auction Market API
  version: "1.0"
paths:
  /auctions:
    get:
      consumes:
      - application/json
      description: Get a paginated list of auctions
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: List auctions
      tags:
      - auctions
    post:
      consumes:
      - application/json
      description: Create a new auction with the provided information
      parameters:
      - description: Auction payload
        in: body
        name: auction
        required: true
        schema:
          $ref: '#/definitions/models.AuctionPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new auction
      tags:
      - auctions
  /auctions/{auctionId}/bids:
    get:
      consumes:
      - application/json
      description: Get all bids for a specific auction
      parameters:
      - description: Auction ID
        in: path
        name: auctionId
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get bids by auction ID
      tags:
      - bids
  /auctions/{id}:
    get:
      consumes:
      - application/json
      description: Get a single auction by ID
      parameters:
      - description: Auction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get auction by ID
      tags:
      - auctions
  /auctions/my:
    get:
      consumes:
      - application/json
      description: Get auctions created by the current user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user auctions
      tags:
      - auctions
...
  /bids/{id}:
    get:
      consumes:
      - application/json
      description: Get a single bid by ID
      parameters:
      - description: Bid ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get bid by ID
      tags:
      - bids
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"